+++ node/src/main/kotlin/net/corda/node/internal/NodeStartup.kt
@@ -27,9 +27,9 @@ import net.corda.node.utilities.registration.UnableToRegisterNodeWithDoormanExce
 import net.corda.node.utilities.saveToKeyStore
 import net.corda.node.utilities.saveToTrustStore
 import net.corda.nodeapi.internal.addShutdownHook
+import net.corda.nodeapi.internal.persistence.DatabaseIncompatibleException
 import net.corda.nodeapi.internal.config.UnknownConfigurationKeysException
 import net.corda.nodeapi.internal.persistence.CouldNotCreateDataSourceException
-import net.corda.nodeapi.internal.persistence.IncompatibleAttachmentsContractsTableName
 import net.corda.tools.shell.InteractiveShell
 import org.fusesource.jansi.Ansi
 import org.fusesource.jansi.AnsiConsole
@@ -174,7 +174,7 @@ open class NodeStartup(val args: Array<String>) {
         } catch (e: NetworkParametersReader.Error) {
             logger.error(e.message)
             return false
-        } catch (e: IncompatibleAttachmentsContractsTableName) {
+        } catch (e: DatabaseIncompatibleException) {
             e.message?.let { Node.printWarning(it) }
             logger.error(e.message)
             return false