+++ docs/source/changelog.rst
@@ -8,6 +8,10 @@ Unreleased
 ----------
 * Introduced ``TestCorDapp`` and utilities to support asymmetric setups for nodes through ``DriverDSL``, ``MockNetwork`` and ``MockServices``.
 
+* Change type of the `checkpoint_value` column. Please check the upgrade-notes on how to update your database.
+
+* Removed buggy :serverNameTablePrefix: configuration.
+
 * ``freeLocalHostAndPort``, ``freePort``, and ``getFreeLocalPorts`` from ``TestUtils`` have been deprecated as they
   don't provide any guarantee the returned port will be available which can result in flaky tests. Use ``PortAllocation.Incremental``
   instead.